requirements.txt
mysql-connector-python==8.2.0
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
boto3==1.34.0
